added user editing + deleting , intgrated harel's new videos, added jwt verification to backend, and jwt handling to front end.

+++ Server/controllers/videoController.js
@@ -1,5 +1,8 @@
 const Video = require('../models/Video');
+const jwt = require("jsonwebtoken")
+const key = "Some super secret key"
 const mongoose = require('mongoose');
+
 
 
 // Create a new video
@@ -12,6 +15,31 @@ exports.createVideo = async (req, res) => {
     res.status(400).send(error);
   }
 };
+
+exports.getVideos = async (req, res) => {
+  try {
+    // Fetch top 10 videos by views
+    const topVideos = await Video.find().sort({ views: -1, _id: 1  }).limit(10);
+    const topVideoIds = topVideos.map(video => video._id);
+
+    // Fetch 10 random videos excluding the top 10
+    const randomVideos = await Video.aggregate([
+      { $match: { _id: { $nin: topVideoIds } } },
+      { $sample: { size: 10 } }
+    ]);
+
+    // Combine both results
+    const videos = [...topVideos, ...randomVideos];
+
+    // Send the combined results
+    res.send(videos);
+  } catch (error) {
+
+    console.error('Error fetching videos:', error);
+    res.status(500).send(error);
+  }
+};
+
 
 exports.getAllVideos = async (req, res) => {
   try {
@@ -84,7 +112,7 @@ exports.getRelatedVideos = async (req, res) => {
         ...pipeline,
         { $skip: skip },
         { $limit: limit },
-        { $sample: { size: limit }}
+        { $sample: { size: limit } }
       ];
     }
 
@@ -103,78 +131,6 @@ exports.getRelatedVideos = async (req, res) => {
     console.log(error);
   }
 };
-
-
-exports.getRelatedVideosWithoutCurrentOnes = async (req, res) => {
-  try {
-    const video = await Video.findById(req.params.id);
-    const { _id, username, tags } = video;
-    const relatedVideosIds = req.body;
-    const objectIdArray = relatedVideosIds
-      .filter(id => {
-        const isValid = mongoose.Types.ObjectId.isValid(id);
-        if (!isValid) {
-          console.error(`Invalid ObjectID: ${id}`);
-        }
-        return isValid;
-      })
-      .map(id => {
-        try {
-          return new mongoose.Types.ObjectId(id);
-        } catch (err) {
-          console.error(`Error converting to ObjectID: ${id}`, err);
-          return null;
-        }
-      })
-      .filter(id => id !== null);
-    // Fetch videos from the database based on the query
-    const pipeline = [
-      {
-        $match: {
-          _id: { $nin: objectIdArray },
-          $or: [
-            { username: username },
-            { tags: { $in: tags } }
-          ]
-        }
-      },
-      {
-        $addFields: {
-          commonTags: {
-            $size: {
-              $setIntersection: [tags, '$tags']
-            }
-          }
-        }
-      },
-      {
-        $sort: { commonTags: -1 } // Sort by most common tags descending
-      },
-      { $limit: 10 },
-      { $sample: { size: 10 } }
-    ];
-
-    const videos = await Video.aggregate(pipeline);
-    if (videos.length < 1) {
-      const pipelineFallback = [
-        {
-          $match: {
-            _id: { $nin: objectIdArray }
-          }
-        },
-        { $sample: { size: 10 } } // Sample 10 random documents
-      ];
-      const fallBackVideos = await Video.aggregate(pipelineFallback);
-      await res.send(fallBackVideos);
-    }
-    else {
-      await res.send(videos);
-    }
-  } catch (error) {
-    res.status(500).send(error);
-  }
-};
-
 // Get a specific video by ID
 exports.getVideoById = async (req, res) => {
   try {
@@ -192,10 +148,20 @@ exports.getVideoById = async (req, res) => {
 // Update a video by ID (PUT)
 exports.updateVideo = async (req, res) => {
   try {
-    const video = await Video.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
-    if (!video) {
+
+
+    const test = await Video.findById(req.params.id);
+    if (!test) {
       return res.status(404).send();
     }
+    const token = req.headers.authorization.split(" ")[1];
+    const data = jwt.verify(token, key);
+    if (data.username !== test.username) {
+      return res.status(403).send('Forbidden');
+    }
+
+    const video = await Video.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+
     res.send(video);
   } catch (error) {
     res.status(400).send(error);
@@ -205,11 +171,71 @@ exports.updateVideo = async (req, res) => {
 // Update a video by ID (PATCH)
 exports.partialUpdateVideo = async (req, res) => {
   try {
-    const video = await Video.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
-    if (!video) {
+    const test = await Video.findById(req.params.id);
+    if (!test) {
       return res.status(404).send();
     }
-    res.send(video);
+    if (req.headers.authorization) {
+      // Extract the token from that header
+      try {
+        const token = req.headers.authorization.split(" ")[1];
+        const data = jwt.verify(token, key);
+        if (data.username !== test.username) {
+          if ('title' in req.body || 'description' in req.body || 'source' in req.body ||
+            'thumbnail' in req.body || 'tags' in req.body || 'upload_date' in req.body ||
+            'duration' in req.body || 'username' in req.body) {
+            return res.status(403).send('Forbidden');
+          }
+
+          if ('views' in req.body) {
+            if (req.body.views !== test.views + 1 && req.body.views !== test.views) {
+              return res.status(403).send('Forbidden');
+            }
+          }
+          if ('likeCount' in req.body) {
+            if (req.body.likeCount !== test.likeCount + 1 && req.body.likeCount !== test.likeCount - 1) {
+              return res.status(403).send('Forbidden');
+            }
+          }
+
+          if ('usersLikes' in req.body) {
+            if (req.body.usersLikes.length !== test.usersLikes.length + 1 && req.body.usersLikes.length !== test.usersLikes.length - 1) {
+              return res.status(403).send('Forbidden');
+            }
+          }
+          if ('comments' in req.body) {
+            if (req.body.comments.length !== test.comments.length + 1 && req.body.comments.length !== test.comments.length - 1) {
+              return res.status(403).send('Forbidden');
+            }
+          }
+        }
+        const video = await Video.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+
+        res.send(video);
+      }
+      catch (err) {
+        if ('title' in req.body || 'description' in req.body || 'source' in req.body ||
+          'thumbnail' in req.body || 'tags' in req.body || 'upload_date' in req.body ||
+          'duration' in req.body || 'username' in req.body || 'likeCount' in req.body ||
+          'usersLikes' in req.body || 'comments' in req.body) {
+          return res.status(403).send('Forbidden');
+        }
+        else if ('views' in req.body) {
+          if (req.body.views !== test.views + 1 && req.body.views !== test.views) {
+            return res.status(403).send('Forbidden');
+          }
+          try {
+            const video = await Video.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+            res.send(video);
+          }
+          catch (error) {
+            res.status(400).send(error);
+          }
+        }
+      }
+    }
+    else
+      return res.status(403).send('Token required');
   } catch (error) {
     res.status(400).send(error);
   }
@@ -218,10 +244,20 @@ exports.partialUpdateVideo = async (req, res) => {
 // Delete a video by ID
 exports.deleteVideo = async (req, res) => {
   try {
-    const video = await Video.findByIdAndDelete(req.params.id);
-    if (!video) {
+
+
+    const test = await Video.findById(req.params.id);
+    if (!test) {
       return res.status(404).send();
     }
+    const token = req.headers.authorization.split(" ")[1];
+    const data = jwt.verify(token, key);
+    if (data.username !== test.username) {
+      return res.status(403).send('Forbidden');
+    }
+
+    const video = await Video.findByIdAndDelete(req.params.id);
+
     res.send(video);
   } catch (error) {
     res.status(500).send(error);
